+++ src/main/java/kr/co/itid/cms/entity/cms/core/menu/Menu.java
@@ -21,10 +21,10 @@ public class Menu {
     private Long parentId;  // 상속메뉴
 
     @Column(name = "position", nullable = false)
-    private Long position;  // 트리 포지션
+    private Integer position;  // 트리 포지션
 
     @Column(name = "level", nullable = false)
-    private Long level;  // 트리 레벨
+    private Integer level;  // 트리 레벨
 
     @Column(name = "title", length = 255)
     private String title;  // 메뉴명
