+++ src/main/java/kr/co/itid/cms/repository/cms/core/content/ContentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ContentRepository extends JpaRepository<Content, Integer> {
+public interface ContentRepository extends JpaRepository<Content, Long> {
 
     /**
      * sort가 0이면서 사용 중인 콘텐츠 목록 (대표 콘텐츠)
@@ -16,16 +16,16 @@ public interface ContentRepository extends JpaRepository<Content, Integer> {
     /**
      * 특정 parentId 그룹의 콘텐츠 목록 (정렬 순서대로)
      */
-    List<Content> findByParentIdOrderBySortAsc(Integer parentId);
+    List<Content> findByParentIdOrderBySortAsc(Long parentId);
 
     /**
      * 특정 parentId에 속한 콘텐츠의 최대 sort값 조회
      */
     @Query("SELECT MAX(c.sort) FROM Content c WHERE c.parentId = :parentId")
-    Integer findMaxSortByParentId(Integer parentId);
+    Integer findMaxSortByParentId(Long parentId);
 
     /**
      * 대표 콘텐츠 및 그 하위 콘텐츠 전체 삭제 (parentId 또는 idx가 일치하는 경우)
      */
-    void deleteAllByParentIdOrIdx(Integer parentId, Integer idx);
+    void deleteAllByParentIdOrIdx(Long parentId, Long idx);
 }
