+++ src/main/java/kr/co/itid/cms/repository/cms/core/permission/PermissionRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface PermissionRepository extends JpaRepository<Permission, Long> {
     @Query("SELECT p FROM Permission p WHERE p.menuId IN :menuIds")
-    List<Permission> findAllByMenuIds(List<Integer> menuIds);
+    List<Permission> findAllByMenuIds(List<Long> menuIds);
 }
